refactor crystals
